Add extra level of abstraction to model
Added Game class that evals all the instructions, places the robot on
the board and prints the report.

The ToyRobot class has been simplified to just execute movements and
change its own position.

The Board class has been changed into interface and SquareBoard
implements it, doing so we can add different boards in the future.